centralize fake dataarray fixtures
